+++ CHANGES.rst
@@ -58,9 +58,18 @@ New Features
 - Add ability to query VO and Astroquery from catalog/using viewer coordinates as the source. The former source selection
   choice 'Manual' is also now 'Source'. [#4275]
 
+- Improvements to spectral axis unit conversion logic to support workflows and
+  avoid errors in mixed unit viewers in deconfigged. [#4292]
+
 - Add Line List Database loader. [#4305, #4314, #4319]
 
+- Add consolidated emission line list and ability to query database in API. [#4314]
+
 - Add ability to view data in multiple viewers with row click in table viewer. [#4279]
+
+- Add ability to specify coordinate frame and equinox in catalog loader. [#4207]
+
+- Table viewer UI to add/rename/remove columns. [#4282]
 
 Mosviz
 ^^^^^^
@@ -80,6 +89,8 @@ Bug Fixes
 - Fixed main widget not displaying in standalone build. [#4271]
 
 - Fixed a Vue3 button issue that triggers backend calls twice. [#4276, #4277]
+
+- Fixed closing a viewer not deleting the associated backend objects. [#4317]
 
 Mosviz
 ^^^^^^
@@ -112,6 +123,8 @@ Bug Fixes
 - Fix image importer support for Roman L3 mosaic files. [#4309]
 
 - Fix issue where snackbar is attached to the notebook rather than to the app. [#4313]
+
+- Allow markers plugin table to handle images with no flux units specified in the header. [#4320]
 
 Mosviz
 ^^^^^^